org-ref multi-citations are handled by "cite:key1,key2". In the org-roam databases this was
previously stored as a single "cite" link to the (non-existing) key "key1,key2". This means that
cite backlinks and cite graph behaviour does not work correctly for these links.

This fix will split any cite links with a comma into separate links to each individual ref.

To be honest, I don't really  like how this patch changes the readability of `org-roam--extract-links`. Because I've basically wrapped the existing code in a `mapccan`, the code to split the multicite links now appears before the code to actually extract the links; in my mind these should be the other way around. In addition, properly indenting would cause a gross diff (I've currently left the original indentation). Not sure how to fix this with my limited emacs-lisp knowledge so suggestions are welcome.

###### Motivation for this change
Fixes #544