@zaeph 2020-05-01
To avoid the long argument in `mapcan`, I've pushed a slight refactoring with `->>`.  Make sure to pull!