@TimQuelch 2020-05-01
> To avoid the long argument in `mapcan`, I've pushed a slight refactoring with `->>`. Make sure to pull!

Oh nice, that's exactly what I wanted. I didn't know about the dash library, it looks very useful