@zaeph 2020-05-01
Welp, the code might be more readable, and I'd certainly deserve some golf-points for it, it isn't that much optimised:
```el
;; Original
(with-selected-window (next-window)
  (benchmark 100 '(org-roam--extract-links)))
;; "Elapsed time: 5.322224s (1.574320s in 17 GCs)"

;; Optimised...?
(with-selected-window (next-window)
  (benchmark 100 '(org-roam--extract-links)))
;; "Elapsed time: 5.352579s (1.565723s in 17 GCs)"
```

I guess it'd be better if we could get `org-element-map` to not accumulate the result, but unless we send a patch upstream (i.e. Org-mode) to get it to accept a new argument, it's not going to happen.

Oh well. ¯\\\_(ツ)\_/¯